requirements.txt
aiofiles
aiohttp
apscheduler
cairosvg
cloudscraper
colour
covid
cowpy
dataclasses
DateTime
emoji
fake_useragent
fonttools
geopy
gitpython
glitch_this
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googletrans==4.0.0-rc1
gtts
hachoir
heroku3
html-telegraph-poster
humanize
IMDbPY
jikanpy
justwatch
lottie
lyricsgenius
markdown
motor
moviepy
nekos.py
Pillow
prettytable
psutil
psycopg2
PyDictionary
pyfiglet
PyGithub
pygments
pylast
pymediainfo
PyMuPDF
pySmartDL
python-barcode
python-dotenv
pytz
qrcode
randomstuff.py
regex
requests
search-engine-parser
selenium
setuptools
ShazamAPI
spamwatch
speedtest-cli
sqlalchemy-json
sqlalchemy==1.3.23
telegraph
git+https://github.com/sandy1709/Telethon
TgCrypto
tswift
ujson
urlextract
validators
vcsi
wand
wget
git+https://github.com/goldsmith/Wikipedia
youtube-search-python
youtube_dl
